Consolidate forward test helpers
