+++ cli/dust-sandbox/src/commands/forward/http2/test_support.rs
@@ -1,4 +1,3 @@
-pub(super) use std::collections::HashMap;
 pub(super) use std::future::{poll_fn, Future};
 pub(super) use std::pin::Pin;
 pub(super) use std::sync::atomic::{AtomicUsize, Ordering};
@@ -14,8 +13,9 @@ pub(super) use tokio::io::{AsyncRead, AsyncReadExt, AsyncWriteExt};
 pub(super) use tokio::sync::mpsc;
 pub(super) use tracing::warn;
 
-pub(super) use crate::egress_secrets::{DomainSet, Secret, SecretTable};
+pub(super) use crate::egress_secrets::SecretTable;
 
+pub(super) use super::super::test_support::{read_h2_body, secret_table_with_secret};
 pub(super) use super::pool::{H2UpstreamPool, UpstreamLease};
 pub(super) use super::{
     run_h2_to_h1_bridge, run_h2_to_upstream_bridge, send_data, BoxedAsyncReadWrite, H2UpstreamKey,
@@ -822,17 +822,6 @@ where
     }
 }
 
-pub(super) async fn read_h2_body(mut body: RecvStream) -> Result<Vec<u8>> {
-    let mut output = Vec::new();
-    while let Some(chunk) = body.data().await {
-        let chunk = chunk?;
-        output.extend_from_slice(&chunk);
-        body.flow_control().release_capacity(chunk.len())?;
-    }
-    ensure!(body.trailers().await?.is_none(), "unexpected trailers");
-    Ok(output)
-}
-
 pub(super) async fn next_informational(
     response: &mut h2::client::ResponseFuture,
 ) -> Result<Response<()>> {
@@ -895,29 +884,4 @@ pub(super) fn assert_h2_reset_reason(error: h2::Error, reason: h2::Reason) {
         Some(reason),
         "unexpected h2 reset reason: {error}"
     );
-}
-
-pub(super) fn secret_table_with_secret(
-    name: &str,
-    placeholder: &str,
-    value: &str,
-    patterns: &[&str],
-) -> Result<SecretTable> {
-    let allowed_domains = patterns
-        .iter()
-        .map(|pattern| (*pattern).to_string())
-        .collect::<Vec<_>>();
-    let domain_set = DomainSet::from_patterns(&allowed_domains)?;
-    let secret = Secret {
-        name: name.to_string(),
-        placeholder: placeholder.to_string(),
-        value: value.to_string(),
-        allowed_domains: domain_set,
-    };
-    let mut by_placeholder = HashMap::new();
-    by_placeholder.insert(placeholder.to_string(), secret);
-    Ok(SecretTable {
-        by_placeholder,
-        sni_match_set: DomainSet::from_patterns(&allowed_domains)?,
-    })
 }
